[embed] Extract ModalLoadingOverlay into a separate component
